+++ lib/appmetrica_push_plugin.dart
@@ -2,6 +2,8 @@ library appmetrica_push;
 
 export 'src/appmetrica_push.dart' show AppMetricaPush;
 
+export 'src/appmetrica_push_info.dart' show AppMetricaPushInfo;
+
 export 'src/push_provider.dart' show PushProvider;
 
 export 'src/firebase_push_provider.dart' show FirebasePushProvider;